% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/output_contains.R
\name{output_contains2}
\alias{output_contains2}
\title{Other version of output_contains useful in some cases}
\usage{
output_contains2(expr, console_output = DM.console.output)
}
\arguments{
\item{expr}{The expression (as string) for which the output should be in the student's console.}

\item{console_output}{The string containing the output printed to the student's console.
The default is \code{DM.console.output} which is set on the DataCamp server (automagically).
This means you don't have to specify this argument when writing SCTs for DataCamp.com.}
}
\description{
Other version of output_contains useful in some cases
}

